+++ .env
@@ -1,2 +1,1 @@
 AUTH_SECRET=6716884e63faa2aac0da9995e28186f6bff5cbaf5e43d18e8b5e1d375c4a6ee34f73de4e92a2c4135f780f565dff12c362af88623af9a61de116c74d956732d9
-PASSWORD='EspertoFitGym'
