+++ app/controllers/units_controller.rb
@@ -12,4 +12,6 @@ class UnitsController < ApplicationController
   def show
     @unit = Unit.find(params[:id])
   end
+
+
 end
